+++ tests/api-resources/endpoints.test.ts
@@ -72,12 +72,26 @@ describe('resource endpoints', () => {
   test('list: request options and params are passed correctly', async () => {
     // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
     await expect(
-      client.endpoints.list({ type: 'dedicated' }, { path: '/_stainless_unknown_path' }),
+      client.endpoints.list(
+        { mine: true, type: 'dedicated', usage_type: 'on-demand' },
+        { path: '/_stainless_unknown_path' },
+      ),
     ).rejects.toThrow(Together.NotFoundError);
   });
 
   test('delete', async () => {
     const responsePromise = client.endpoints.delete('endpoint-d23901de-ef8f-44bf-b3e7-de9c1ca8f2d7');
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  test('listAvzones', async () => {
+    const responsePromise = client.endpoints.listAvzones();
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;